requirements.txt
twitchio==2.6.0
aiohttp==3.8.4
websockets==10.4
python-dotenv==1.0.0
openai==0.27.0
ratelimit==2.2.1
packaging
pyaudio
requests
setuptools
flake8==3.9.2
black==21.6b0